requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cloud-sql-python-connector==1.12.0
cryptography==43.0.1
frozenlist==1.4.1
google-auth==2.34.0
greenlet==3.0.3
idna==3.8
multidict==6.0.5
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyMySQL==1.1.1
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
soupsieve==2.6
SQLAlchemy==2.0.34
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.10.0
